Fixes - Only Tinkertools have the repair context mode - Fixed Gary from being spawned on server start over and over!
