+++ Scripts/Items/Tools/BaseTool.cs
@@ -302,40 +302,5 @@ namespace Server.Items
             return quality;
         }
         #endregion
-
-        public override void GetContextMenuEntries(Mobile from, List<ContextMenuEntry> list)
-        {
-            base.GetContextMenuEntries(from, list);
-
-            if (Core.TOL)
-                list.Add(new ToggleRepairContextMenuEntry(from, this));
-        }
-
-        public class ToggleRepairContextMenuEntry : ContextMenuEntry
-        {
-            private Mobile _From;
-            private BaseTool _Tool;
-
-            public ToggleRepairContextMenuEntry(Mobile from, BaseTool tool)
-                : base(1157040) // Toggle Repair Mode
-            {
-                _From = from;
-                _Tool = tool;
-            }
-
-            public override void OnClick()
-            {
-                if (_Tool.RepairMode)
-                {
-                    _From.SendLocalizedMessage(1157042); // This tool is fully functional. 
-                    _Tool.RepairMode = false;
-                }
-                else
-                {
-                    _From.SendLocalizedMessage(1157041); // This tool will only repair items in this mode.
-                    _Tool.RepairMode = true;
-                }
-            }
-        }
     }
 }
